+++ integration/app/app.browser.module.ts
@@ -3,7 +3,7 @@ import { StateContext } from '@ngxs/store';
 import { BrowserTransferStateModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgxsStoragePluginModule } from '@ngxs/storage-plugin';
-import { NgxsHmrLifeCycle, NgxsStoreSnapshot as Snapshot } from '@ngxs/hmr-plugin';
+import { NgxsHmrLifeCycle, NgxsHmrSnapshot as Snapshot } from '@ngxs/hmr-plugin';
 
 import { AppComponent } from '@integration/app.component';
 import { AppModule } from '@integration/app.module';
@@ -20,15 +20,11 @@ import { TODOS_STORAGE_KEY } from '@integration/store/todos/todos.model';
   providers: [{ provide: 'ORIGIN_URL', useValue: location.origin }]
 })
 export class AppBrowserModule implements NgxsHmrLifeCycle<Snapshot> {
-  public hmrNgxsStoreOnInit(ctx: StateContext<Snapshot>, snapshot: Snapshot) {
-    console.log('[NGXS HMR] Current state', ctx.getState());
-    console.log('[NGXS HMR] Previous state', snapshot);
+  public hmrNgxsStoreOnInit(ctx: StateContext<Snapshot>, snapshot: Partial<Snapshot>) {
     ctx.patchState(snapshot);
   }
 
-  public hmrNgxsStoreBeforeOnDestroy(ctx: StateContext<Snapshot>): Snapshot {
-    const snapshot: Snapshot = ctx.getState();
-    console.log('[NGXS HMR] Saved state before on destroy', snapshot);
-    return snapshot;
+  public hmrNgxsStoreBeforeOnDestroy(ctx: StateContext<Snapshot>): Partial<Snapshot> {
+    return ctx.getState();
   }
 }