+++ integration/app/detail/detail.component.ts
@@ -1,7 +1,13 @@
 import { Component } from '@angular/core';
+import { Select } from '@ngxs/store';
+import { Observable } from 'rxjs';
+
+import { DetailState } from '@integration/detail/detail.state';
 
 @Component({
   selector: 'app-detail',
   templateUrl: './detail.component.html'
 })
-export class DetailComponent {}
+export class DetailComponent {
+  @Select(DetailState) detail$: Observable<boolean>;
+}