+++ packages/store/src/symbols.ts
@@ -1,7 +1,9 @@
-import { Injectable, InjectionToken } from '@angular/core';
+import { Injectable, InjectionToken, Type } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { ObjectKeyMap } from './internal/internals';
+import { NgxsExecutionStrategy } from './execution/symbols';
+import { DispatchOutsideZoneNgxsExecutionStrategy } from './execution/dispatchOutsideZoneNgxsExecutionStrategy';
 
 export const ROOT_STATE_TOKEN = new InjectionToken<any>('ROOT_STATE_TOKEN');
 export const FEATURE_STATE_TOKEN = new InjectionToken<any>('FEATURE_STATE_TOKEN');
@@ -36,12 +38,17 @@ export class NgxsConfig {
     strictContentSecurityPolicy: boolean;
   };
   /**
-   * Determines whether to perform async operations inside or outside Angular's zone, every `zone.run`
-   * causes Angular to run change detection on the whole tree (`app.tick()`). P.S. this will not break
-   * server side rendering, as dispatching on the server is still wrapped into `zone.run`.
+   * Determines the execution context to perform async operations inside. An implementation can be
+   * provided to override the default behaviour where the async operations are run
+   * outside Angular's zone but all observable behaviours of NGXS are run back inside Angular's zone.
+   * These observable behaviours are from:
+   *   `@Select(...)`, `store.select(...)`, `actions.subscribe(...)` or `store.dispatch(...).subscribe(...)`
+   * Every `zone.run` causes Angular to run change detection on the whole tree (`app.tick()`) so of your
+   * application doesn't rely on zone.js running change detection then you can switch to the
+   * `NoopNgxsExecutionStrategy` that doesn't interact with zones.
    * (default: null)
    */
-  outsideZone: boolean | null = null;
+  executionStrategy: Type<NgxsExecutionStrategy>;
   /**
    * Defining the default state before module initialization
    * This is convenient if we need to create a define our own set of states.
@@ -53,6 +60,7 @@ export class NgxsConfig {
     this.compatibility = {
       strictContentSecurityPolicy: false
     };
+    this.executionStrategy = DispatchOutsideZoneNgxsExecutionStrategy;
   }
 }