+++ CHANGELOG.md
@@ -10,10 +10,15 @@ $ npm install @ngxs/store@dev
 * Feature: Add `Hot Module Replacement` plugin for manage states [#707](https://github.com/ngxs/store/pull/707)
 * Feature: Add possible inheritance of state options [#750](https://github.com/ngxs/store/pull/750)
 * Feature: Add new lifecycle hook `ngxsAfterBootstrap` [#753](https://github.com/ngxs/store/pull/753)
-* Feature: Add `outsideZone` option to the config [#761](https://github.com/ngxs/store/pull/761)
-* Feature: Add state operators: `patch, updateItem, removeItem, insertItem, append, compose` [#799](https://github.com/ngxs/store/pull/799/)
+* Feature: Add ability to change the Execution strategy (decouple from zone.js) [#811](https://github.com/ngxs/store/pull/811)
+* Feature: Add state operators: `patch, updateItem, removeItem, insertItem, append, compose, iif` [#799](https://github.com/ngxs/store/pull/799/)
+* Feature: WebSocket Plugin - Add `WebSocketDisconnected` action to notify of disconnection [#825](https://github.com/ngxs/store/pull/825)
+* Feature: Defining the default state before module initialization [#791](https://github.com/ngxs/store/pull/791/files#diff-fc35f12bac15f5f65bad0f323be7ae12R46)
 * Fix: Expose `ActionCompletion` [#752](https://github.com/ngxs/store/pull/752)
 * Fix: Throw error when found duplicate state names [#791](https://github.com/ngxs/store/pull/791)
+* Fix: Bind static context to the selector function [#818](https://github.com/ngxs/store/pull/818)
+* Fix: WebSocket Plugin - `WebsocketMessageError` notifies of errors [#825](https://github.com/ngxs/store/pull/825)
+* Fix: Performance improved reading the name of the state from the parameter [#826](https://github.com/ngxs/store/pull/826)
 
 ## NGXS-Labs